modularizes counterSlice
