+++ src/store/index.js
@@ -1,31 +1,6 @@
 import { createSlice, configureStore } from '@reduxjs/toolkit';
 
-const initialCounterState = { counter: 0, showCounter: true };
 
-// takes object as argument
-// each slice needs a name/identifier
-// redux toolkit uses immer to detect state changes
-// can change aspects of state we want to change
-// state is still immutable
-// immer updates state in the background
-const counterSlice = createSlice({
-    name: 'counter',
-    initialState: initialCounterState,
-    reducers: {
-        increment(state) {
-            state.counter++;
-        },
-        decrement(state) {
-            state.counter--;
-        },
-        increase(state, action) {
-            state.counter = state.counter + action.payload;
-        },
-        toggleCounter(state) {
-            state.showCounter = !state.showCounter;
-        },
-    },
-});
 
 const initialAuthState = {
     isAuthenticated: false,
